Move data migrations out into shared and rename to migrateDataInBatches
