+++ packages/shared/src/dataMigrations/index.js
@@ -1,0 +1,2 @@
+export { DataMigration } from './DataMigration';
+export { CursorDataMigration } from './CursorDataMigration';
