--- packages/sync-server/app/subCommands/migrateData/index.js
@@ -1,1 +1,0 @@
-export { migrateDataCommand } from './migrateData';
