+++ packages/sync-server/app/subCommands/migrateDataInBatches/index.js
@@ -1,0 +1,1 @@
+export { migrateDataInBatchesCommand } from './migrateDataInBatches';
